requirements.txt
asgiref==3.2.5
asn1crypto==1.3.0
certifi==2020.4.5.1
cffi==1.14.0
chardet==3.0.4
Django==3.0.4
django-cors-headers==3.2.1
django-crontab==0.7.1
djangorestframework==3.11.0
httpie==2.1.0
idna==2.9
Pillow==7.1.2
pycparser==2.20
Pygments==2.6.1
PyNaCl==1.4.0
pytz==2019.3
requests==2.23.0
six==1.15.0
sqlparse==0.3.1
urllib3==1.25.9
